Extract model into superclass
